Upgrade network fails: "The IP Address 10.1.0.2 is currently in use"
If the user creates a network using a network offering without VR,
and deploys a VM, specifying the IP the VR would normally use,
upgrading to a network offering  with VR will fail to spin the VR,
as its IP is in use.

We now block the usage of the VR IP, even for non-VR offerings.

BUG-ID: CLOUD-635
Bugfix-for: 4.5
Reviewed-by:
Reported-by: Shiv Gurunathan <shiv.gurunathan@nuagenetworks.net>
Signed-off-by: Frank Maximus <frank.maximus@nuagenetworks.net>